move level class to entity directory
